fix util functions

+++ BlissGBA/Utils.h
@@ -12,8 +12,8 @@ using s16 = int16_t;
 using u32 = uint32_t;
 using s32 = int32_t;
 
-u8 setBit(u8 val, u8 bitToSet);
-u8 resetBit(u8 val, u8 bitToReset);
+u32 setBit(u32 val, u8 bitToSet);
+u32 resetBit(u32 val, u8 bitToReset);
 u8 testBit(u8 val, u8 bitToTest);
 
 //Return value within bit range of from - to
